Added Footer.tsx to Components and Seperated the footer from the HomeScreen.tsx so it can be used elsewhere
Added function find_key() to Footer.tsx to provide an easy way to find the API key on other pages where the key is needed
