+++ src/Pages/HomeScreen.tsx
@@ -1,27 +1,9 @@
-import React, {useState} from "react";
-import {Button, Form} from "react-bootstrap";
+import React from "react";
 import { Link } from "react-router-dom";
 import "../CSS/HomeScreen.css";
 import { NavB } from "../Components/navBar";
-
-
-//local storage and API Key: key should be entered in by the user and will be stored in local storage (NOT session storage)
-let keyData = "";
-const saveKeyData = "MYKEY";
-const prevKey = localStorage.getItem(saveKeyData); //so it'll look like: MYKEY: <api_key_value here> in the local storage when you inspect
-if (prevKey !== null) {
-  keyData = JSON.parse(prevKey);
-}
-
-
-function Header() {
-    /**
-     * JSX element that renders a header that appears at the top of all pages.
-     */
-    return (
-    <NavB/>
-  );
-}
+import "../Components/Footer";
+import Footer from "../Components/Footer";
 
 
 function TestBlurb({header, textBody, linkText, link}: {header: string; textBody: string; linkText: string; link: string;}): JSX.Element {
@@ -43,27 +25,17 @@ function TestBlurb({header, textBody, linkText, link}: {header: string; textBody
     )
 }
 
+
 export default function HomeScreen() {
-  const [key, setKey] = useState<string>(keyData); //for api key input
-
-  //sets the local storage item to the api key the user inputted
-  function handleSubmit() {
-    localStorage.setItem(saveKeyData, JSON.stringify(key));
-    window.location.reload();//when making a mistake and changing the key again, I found that I have to reload the whole site before openai refreshes what it has stores for the local storage variable
-  }
-
   //whenever there's a change it'll store the api key in a local state called key but it won't be set in the local storage until the user clicks the submit button
-  function changeKey(event: React.ChangeEvent<HTMLInputElement>) {
-    setKey(event.target.value);
-  }
   return (
     <div>
-      <Header/>
+      <NavB/>
       <div className="Home-body">
         <TestBlurb
             header={"Basic Test"}
             textBody={
-                "Concerned about the path of your career? Gaining insight into your career option " +
+                "Concerned about the path of your career? Gaining insight into your career options" +
                 "is quick and simple with our basic career test. It provides individualized recommendations " +
                 "based on your interests and strengths by utilizing AI analysis. Easily learn about potential " +
                 "job choices and gain a better understanding of your future in the workforce. Our basic career " +
@@ -88,18 +60,7 @@ export default function HomeScreen() {
             link={"/detailed-test"}
         />
       </div>
-      <Form>
-        <Form.Label>API Key:</Form.Label>
-        <Form.Control
-          type="password"
-          placeholder="Insert API Key Here"
-          onChange={changeKey}
-        ></Form.Control>
-        <br></br>
-        <Button className="Submit-Button" onClick={handleSubmit}>
-          Submit
-        </Button>
-      </Form>
+        <Footer/>
     </div>
   );
 }
